extract formatTime
